+++ src/fapilog/container.py
@@ -14,8 +14,8 @@ import structlog
 
 from ._internal.metrics import MetricsCollector
 from ._internal.processor_metrics import ProcessorMetrics
-from ._internal.queue_worker import QueueWorker
 from .async_components.concurrency.lock_manager import ProcessorLockManager
+from .async_components.queue.worker import QueueWorker
 from .core.factories.component_factory import ComponentFactory
 from .core.factories.container_logger_factory import ContainerLoggerFactory
 from .core.managers.configuration_manager import ConfigurationManager
